requirements.txt
pyyaml>=5.3.1,<6
requests>=2.25.1
backoff>=1.10.0
dohq-artifactory>=0.7.574
docker>=5.0.0
docker-compose>=1.29.2
